// src/types/types.d.ts
interface UserResponse {
  questionId: string;
  response: any;
}
interface ProgressInfo {
  userTags: string[];
  userProgress: UserProgress;
  questionData: QuestionList[];
}
interface QuestionList extends WithId<Document> {
  _id?: ObjectId;
  name: string;
  isCheckin?: boolean;
  part?: string;
  order?: number;
  chunks: QuestionChunk[];
}
interface QuestionList_Db extends WithId<Document> {
  _id?: ObjectId;
  name: string;
  isCheckin?: boolean;
  part?: string;
  order?: number;
  chunks: string[]; // Document IDs of chunks
}
interface QuestionChunk extends WithId<Document> {
  _id?: ObjectId;
  name: string;
  questions: Question[];
}
interface QuestionChunk_Db extends WithId<Document> {
  _id?: ObjectId;
  name: string;
  questions: ObjectId[]; // Document IDs of question data
}
interface Question extends WithId<Document> {
  _id?: ObjectId;
  question: string;
  type: string;
  helpVid?: string;
  helpText?: string;
  data?: any[];
  isConcatenable?: boolean;
}